refactor(logs): inline SentryLog.Create instead of a separate partial file
Move the Create factory method into SentryLog.cs, drop the now-unneeded
partial modifier, and remove the SentryLog.Factory.cs DependentUpon entry
from the csproj. A single method doesn't warrant its own file.
