+++ src/Sentry/Sentry.csproj
@@ -194,9 +194,6 @@
     <Compile Update="MeasurementUnit.Duration.cs;MeasurementUnit.Fraction.cs;MeasurementUnit.Information.cs">
       <DependentUpon>MeasurementUnit.cs</DependentUpon>
     </Compile>
-    <Compile Update="SentryLog.Factory.cs">
-      <DependentUpon>SentryLog.cs</DependentUpon>
-    </Compile>
     <Compile Update="SentryMetric.Factory.cs;SentryMetric.Generic.cs">
       <DependentUpon>SentryMetric.cs</DependentUpon>
     </Compile>
